# pos-service/openapi.yaml
openapi: 3.0.3
info:
  title: POS Service API
  version: 1.0.0
  description: API untuk mencatat transaksi penjualan dan menerbitkan event sale.created.
servers:
  - url: http://localhost:3001
paths:
  /health:
    get:
      summary: Health check POS Service
      responses:
        "200":
          description: Service aktif
  /sales:
    post:
      summary: Membuat transaksi penjualan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSaleRequest"
            example:
              buyerName: Budi
              paymentMethod: cash
              items:
                - productId: P001
                  name: Keyboard
                  quantity: 2
                  price: 150000
    responses:
      "201":
        description: Transaksi dibuat dan event sale.created dipublikasikan
      "400":
        description: Request tidak valid
    get:
      summary: Melihat semua transaksi penjualan
      responses:
        "200":
          description: Daftar transaksi
  /sales/{id}:
    get:
      summary: Melihat detail transaksi berdasarkan ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: SALE-1717322400000
      responses:
        "200":
          description: Detail transaksi
        "404":
          description: Transaksi tidak ditemukan
components:
  schemas:
    CreateSaleRequest:
      type: object
      required:
        - buyerName
        - paymentMethod
        - items
      properties:
        buyerName:
          type: string
          example: Budi
        paymentMethod:
          type: string
          example: cash
        items:
          type: array
          minItems: 1
          items:
            type: object
            required:
              - productId
              - name
              - quantity
              - price
            properties:
              productId:
                type: string
                example: P001
              name:
                type: string
                example: Keyboard
              quantity:
                type: integer
                example: 2
              price:
                type: integer
                example: 150000